+++ SwiftMessages/TopBottomAnimation.swift
@@ -67,18 +67,41 @@ public class TopBottomAnimation: NSObject, Animator {
         }
         container.addConstraints([leading, trailing, translationConstraint])
         if let adjustable = view as? MarginAdjustable {
-            var top: CGFloat = 0.0
-            var bottom: CGFloat = 0.0
+            // Important to layout now in order to get the right safe area insets
+            container.layoutIfNeeded()
+            var top: CGFloat = 0
+            var bottom: CGFloat = 0
             switch style {
             case .top:
                 top += adjustable.bounceAnimationOffset
-                if context.behindStatusBar {
-                    top += adjustable.statusBarOffset
+                if !context.safeZoneConflicts.isDisjoint(with: [.sensorNotch, .statusBar]) {
+                    if #available(iOS 11, *), container.safeAreaInsets.top > 0  {
+                        // Linear formula based on:
+                        // iPhone 8 - 20pt top safe area with 0pt adjustment
+                        // iPhone X - 44pt top safe area with -6pt adjustment
+                        top -= 6 * (container.safeAreaInsets.top - 20) / (44 - 20)
+                        top += adjustable.safeAreaTopOffset
+                    } else {
+                        top += adjustable.statusBarOffset
+                    }
+                }
+                if #available(iOS 11, *), !context.safeZoneConflicts.isDisjoint(with: .coveredStatusBar) {
+                    top -= view.safeAreaInsets.top
                 }
             case .bottom:
                 bottom += adjustable.bounceAnimationOffset
+                if !context.safeZoneConflicts.isDisjoint(with: [.homeIndicator]) {
+                    if #available(iOS 11, *), container.safeAreaInsets.bottom > 0  {
+                        bottom += adjustable.safeAreaBottomOffset
+                    }
+                }
             }
-            view.layoutMargins = UIEdgeInsets(top: top, left: 0.0, bottom: bottom, right: 0.0)
+            view.preservesSuperviewLayoutMargins = false
+            if #available(iOS 11, *) {
+                view.directionalLayoutMargins = NSDirectionalEdgeInsets(top: top, leading: view.directionalLayoutMargins.leading, bottom: bottom, trailing: view.directionalLayoutMargins.trailing)
+            } else {
+                view.layoutMargins = UIEdgeInsets(top: top, left: view.layoutMargins.left, bottom: bottom, right: view.layoutMargins.right)
+            }
         }
         let size = view.systemLayoutSizeFitting(UILayoutFittingCompressedSize)
         translationConstraint.constant -= size.height
@@ -93,6 +116,14 @@ public class TopBottomAnimation: NSObject, Animator {
             }
         }
     }
+
+//    @available(iOS 11, *)
+//    private func adjustForCoveredStatusBarSafeArea(view: UIView) {
+//        var margins = view.directionalLayoutMargins
+//        margins.top -= 2 * view.safeAreaInsets.top
+//        view.directionalLayoutMargins = margins
+//        view.subviews.forEach { self.adjustForCoveredStatusBarSafeArea(view: $0) }
+//    }
 
     func showAnimation(completion: @escaping AnimationCompletion) {
         guard let container = containerView else {