+++ CHANGELOG.md
@@ -1,6 +1,30 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 4.0.0
+
+### Features
+* Swift 4.0 syntax
+* Added support for iOS 11 and iPhone X. From the readme:
+
+  SwiftMessages 4 supports iOS 11 out-of-the-box with built-in support   for safe areas. To ensur  that message view layouts look just right when overlapping safe areas, views that adopt the `MarginAdjustable` protocol (like `MessageView`) will have their layout margins automatically adjusted by SwiftMessages. However, there is no one-size-fits-all adjustment, so the following properties were added to `MarginAdjustable` to allow for additional adjustments to be made to the layout margins:
+
+  ````swift
+public protocol MarginAdjustable {
+    ...
+    /// Safe area top adjustment in iOS 11+
+    var safeAreaTopOffset: CGFloat { get set }
+    /// Safe area bottom adjustment in iOS 11+
+    var safeAreaBottomOffset: CGFloat { get set }
+}
+````
+
+  If you're using using custom nib files or view classes and your layouts don't look quite right, try adjusting the values of these properties. `BaseView` (the super class of `MessageView`) declares these properties to be `@IBDesignable` and you can find sample values in the nib files included with SwiftMessages.
+
+### Bug Fixes
+* Fix #100 memory leak.
+* Change `Layout` enum capitalization to current Swift conventions.
+
 ## [3.5.1](https://github.com/SwiftKickMobile/SwiftMessages/releases/tag/3.5.0)
 
 ### Bug Fixes